@@ -14,17 +14,26 @@ Gem::Specification.new do |s|
   s.rubyforge_project = "blacklight"
 
   s.files         = `git ls-files`.split("\n")
-  s.test_files    = `git ls-files -- {test,spec,features}/*`.split("\n")
+  s.test_files    = Dir["spec/**/*"]
   s.executables   = `git ls-files -- bin/*`.split("\n").map{ |f| File.basename(f) }
   s.require_paths = ["lib"]
 
 
-  s.add_dependency "rails", "~> 4.0"
-  s.add_dependency "blacklight", "~> 6.0"
-  s.add_dependency "oai"
+  s.add_dependency 'rails', '>= 4.2'
+  s.add_dependency 'blacklight', '>= 6.1'
+  s.add_dependency 'oai'
+
   s.add_development_dependency 'rspec'
-  s.add_development_dependency 'rspec-rails'
+  s.add_development_dependency 'rspec-rails', '~> 3.5'
   s.add_development_dependency 'capybara'
+  s.add_development_dependency 'poltergeist'
+  s.add_development_dependency 'sqlite3'
+  s.add_development_dependency 'solr_wrapper', '>= 0.19'
+  s.add_development_dependency 'rsolr', '~> 1.0'
+  s.add_development_dependency 'blacklight-marc', '~> 6.1'
+  s.add_development_dependency 'sass-rails', '~> 5.0'
+  s.add_development_dependency 'jquery-rails'
+  s.add_development_dependency 'byebug'
   s.add_development_dependency 'vcr'
-  s.add_development_dependency 'fakeweb'
+  s.add_development_dependency 'webmock'
 end
